Dipole game on verilog through vga for fpga
This is the android game named dipole. The code is built on verilog
language for fpga.

It can be imlemented directly to altera fpga boards because they are
project files.

There is movement on circle via Newton–Raphson method to find square
root of circle equation then find Y-axis position.